+++ Old_Codes/cifras_functions.py
