+++ Old_Codes/cripto_hash.py
