+++ Old_Codes/cripto_modern.py
