+++ Old_Codes/cripto_simetrica_imagens.py
